Move helper methods out into own db esque file
